+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`633` When using the `handling_exceptions`, it is now possible to obtain the exception object that was handled
 * :feature:`-` Added ``SLASH_USER_SETTINGS=x`` environment variable to give a possibility to override the user slashrc file
 * :feature:`592` Added ``exception_attributes`` dict to ``Error`` objects
 * :feature:`600` Use `vintage` package for deprecations